Perf update on pathing
MM was checking if 2 actors where the same via a UniqueID check, however string equality is pretty slow.

Instead UniqueIDs are now mapped to an integer ID and actors are then mapped using this to an Actor to Integer ID, allowing for quick dictionary lookups and integer comparison to see if they are the same unit.

This should speed up HasCollisionAt